adding ft_putnbr, ft_strcmp and ft_strdup

+++ srcs/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: amann <amann@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 12:56:09 by amann             #+#    #+#             */
-/*   Updated: 2021/11/01 14:46:58 by amann            ###   ########.fr       */
+/*   Updated: 2021/11/01 15:23:09 by amann            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(char *str)
+size_t ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
